Added support for Mesos-DNS
Following the installation of Marathon, Mesos-DNS packages are pull on each slave node.
A sample mesos-dns config.json file is issued and configured with the proper ZK address and resolvers.

Mesos-DNS should afterward be run with Marathon on one or more slaves. The /etc/resolv.conf file must be updated on each slave accordingly.To allow service discovery using Marathon-lb, we also added reserved ports in the slaves' configuration.

NOTE: This is WIP and is actually designed for a POC. In the future, we should distinguish slaves set up with service discovery features from the other slaves in the cluster.